Split album route files
