+++ app/routes/api/api.router.js
@@ -2,7 +2,7 @@ const path = require("path");
 const router = require("express").Router();
 const rateLimit = require("express-rate-limit");
 const passport = require("passport");
-const albumRouter = require("./album.api.router");
+const albumRouter = require("./album/router");
 const artistRouter = require("./artist.api.router");
 const crateRouter = require("./crate.api.router");
 const playlistRouter = require("./playlist.api.router");
